# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ecommerce Api",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "project2-d49k.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
		"/shops/{shopId}/products": {
			"get": {
				"tags": ["Shops"],
				"description": "",
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "shopId",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
...
		"/shops/{shopId}/categories": {
			"get": {
				"tags": ["Shops"],
				"description": "",
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "shopId",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
...
	}
}